backend/requirements.txt
fastapi 
uvicorn 
python-dotenv
websockets
python-multipart
aiohttp
genai
google-genai>=1.16.0
pillow
mss
googletrans
google-adk==1.8.0
certifi
googlemaps
geopy
google-generativeai


